--- CNAME
@@ -1,1 +1,0 @@
-hvltech.de
