+++ hvlMeetup/src/App.css
@@ -5,7 +5,7 @@
 }
 
 body,html{
-  height: 100vh;
+  height: 100%;
   width: 100%;
   color: #00274a;
   font-size: 18px;
@@ -13,7 +13,14 @@ body,html{
   font-family: 'Iosevka', monospace, 'Roboto Light', 'Roboto',  Helvetica, Arial, sans-serif;
 
 }
-
+#root{
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+main{
+  flex:1;
+}
 
 h1{
   font-family: "Press Start 2P", system-ui;
@@ -22,7 +29,7 @@ h1{
   font-size: 1.8rem;
   text-align: center;
   text-shadow: 0px 0px 10px white ;
-  margin-top: 50px;
+
 }
 
 #head-event{
@@ -33,8 +40,9 @@ h1{
   height: 1000px;
   max-height: 598px;
   width: 100%;
+
 }
-.translate-block{
+.nav{
   display: flex;
   gap:10px;
   justify-content: end;
@@ -63,8 +71,8 @@ h1{
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: space-between;
-  height: 100%;
+  justify-content: space-around;
+  height: 92%;
   width: 100%;
 }
 
@@ -143,7 +151,6 @@ h1{
 }
 
 .main{
-  flex:1;
   background:white;
   gap: 10px;
   text-align:justify ;
@@ -209,11 +216,13 @@ footer{
     display: flex;
     margin: 0 auto;
 
+
   }
   #head-event{
     width: 1120px;
     margin: 0 auto;
   }
+
 }
