+++ hvlMeetup/src/main.jsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
 
-import './i18n'
+import './utils/i18n.js'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
