+++ hvlMeetup/src/translate/de.js
@@ -1,0 +1,30 @@
+const de= {
+  translation: {
+    countdown: {
+      days: 'Tagen',
+      hours: 'Stunden',
+      minutes: 'Minuten',
+      seconds: 'Sekunden'
+    },
+    buttonJoinUs: 'sei dabei!',
+    aboutAs: {
+      highline: 'Über uns',
+      description: '<0>Willkommen bei <1> Havelland Technology Meetup</1> der führenden Meetup-Gruppe für Softwareentwickler*innen in Falkensee und Umgebung!\n'
+          +
+          '  Wir sind überzeugt, dass in Falkensee einige der besten Entwicklerinnen zu Hause sind, und laden alle Technikbegeisterten herzlich ein, mitzumachen.\n'
+          +
+          ' Egal, mit welcher Sprache oder Plattform du arbeitest – bei uns ist jeder willkommen.</0>'
+    },
+    whyTakePart: {
+      highline: 'Warum mitmachen?',
+      benefits1: 'Kostenlose Teilnahme, offen für alle Technologien und Erfahrungsstufen',
+      benefits2: 'Bau dein lokales Netzwerk auf – freundliche und entspannte Atmosphäre',
+      benefits3: 'Entdecke neue Technologien',
+      benefits4: 'Lass dich von Vorträgen, Demos und Gesprächen inspirieren',
+      benefits5: 'Vernetze dich mit Entwickler*innen aus der Region'
+    }
+  }
+}
+
+
+export default de;
