--- hvlMeetup/src/translate/de.json
@@ -1,4 +1,0 @@
-
-
-
-export
