+++ hvlMeetup/src/translate/en.js
@@ -1,0 +1,30 @@
+const en= {
+        translation: {
+            countdown: {
+                days: 'days',
+                hours: 'hours',
+                minutes: 'minutes',
+                seconds: 'seconds',
+            },
+            buttonJoinUs: 'join us!',
+            aboutAs: {
+                highline: 'About us',
+                description: '<0>Welcome to <1> Havelland Technology Meetup</1> the premier meetup group for software developers\n' +
+                    '                        in\n' +
+                    '                        Falkensee &\n' +
+                    '                        the surrounding area! We believe that Falkensee is home to some of the best developers, and we\n' +
+                    '                        invite all tech enthusiasts to join us. No matter what language or platform you work with,\n' +
+                    '                        everyone is welcome here.</0>'
+            },
+            whyTakePart: {
+                highline: 'Why take part?',
+                benefits1: 'Free to attend, open to all tech stacks and experience levels',
+                benefits2: 'Build your local network – friendly and relaxed atmosphere',
+                benefits3: 'Discover new technologies',
+                benefits4: 'Get inspired by talks, demos, and conversations',
+                benefits5: 'Connect with developers from the region',
+
+            }
+        }
+}
+export default en
