+++ index.html
@@ -17,7 +17,7 @@
 
 <body>
 <header>
-    <div><button id="language"></button>English</div>
+
 </header>
 <main>
   <section id="head-event">
