--- package.json
@@ -1,15 +1,0 @@
-{
-  "name": "hlt-meetup-4",
-  "version": "1.0.0",
-  "description": "hltMeetup",
-  "main": "js/index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "danka",
-  "license": "ISC",
-  "dependencies": {
-    "i18next": "^25.0.1"
-  }
-}
